Added support for rv2.0.0-p247
Why: I am experiencing a similar problem to that describe in the original
project with a long running Ruby application that does file uploads to a
website.  The gem that does the uploading uses tempfile.  I'm running on
the indicated ruby version and patch level so I wanted to see if this
patch would fix my issue.